A super basic website, developed in the mindset that the site is user friendly.
Super basic project created in my Human and Computer Interaction class.
Based off of creating a site that is easy to navigate and overall user
friendly. Much of the site does not include code that allows for login,
or registration as that would have required a server and some PHP
coding, the login page and registration page are pseudo pages meant to
exist if there was an official server for this site.